# Deprecate manually vectorized methods in favor of dot syntax, monolithic edition