--- a/base/floatfuncs.jl
+++ b/base/floatfuncs.jl
@@ -112,29 +112,6 @@ function round(x::AbstractFloat, ::RoundingMode{:NearestTiesUp})
 end
 round{T<:Integer}(::Type{T}, x::AbstractFloat, r::RoundingMode) = trunc(T,round(x,r))
 
-for f in (:ceil,)
-    @eval begin
-        function ($f){T,R}(::Type{T}, x::AbstractArray{R,1})
-            [ ($f)(T, y)::T for y in x ]
-        end
-        function ($f){T,R}(::Type{T}, x::AbstractArray{R,2})
-            [ ($f)(T, x[i,j])::T for i = 1:size(x,1), j = 1:size(x,2) ]
-        end
-        function ($f){T}(::Type{T}, x::AbstractArray)
-            reshape([ ($f)(T, y)::T for y in x ], size(x))
-        end
-        function ($f){R}(x::AbstractArray{R,1}, digits::Integer, base::Integer=10)
-            [ ($f)(y, digits, base) for y in x ]
-        end
-        function ($f){R}(x::AbstractArray{R,2}, digits::Integer, base::Integer=10)
-            [ ($f)(x[i,j], digits, base) for i = 1:size(x,1), j = 1:size(x,2) ]
-        end
-        function ($f)(x::AbstractArray, digits::Integer, base::Integer=10)
-            reshape([ ($f)(y, digits, base) for y in x ], size(x))
-        end
-    end
-end
-
 # adapted from Matlab File Exchange roundsd: http://www.mathworks.com/matlabcentral/fileexchange/26212
 # for round, og is the power of 10 relative to the decimal point
 # for signif, og is the absolute power of 10