… base/linalg/tridiag.jl
… base/math.jl
… base/sparse/higherorderfns.jl
… base/sparse/sparsematrix.jl
… base/sparse/sparsevector.jl
… base/special/trig.jl
… examples/lru_test.jl
… test/bigint.jl
… test/bitarray.jl
… test/floatfuncs.jl
… test/linalg/arnoldi.jl
… test/linalg/bidiag.jl
… test/linalg/generic.jl
… test/linalg/qr.jl
… test/linalg/triangular.jl
… test/linalg/tridiag.jl
… test/math.jl
… test/numbers.jl
… test/offsetarray.jl
… test/random.jl
… test/ranges.jl
… test/reduce.jl
… test/sparse/sparse.jl
… test/sparse/sparsevector.jl