@tkelman 2016-12-31
This will likely show big regressions for any of the benchmarks that are using now-deprecated syntax. I'm more curious about whether it shows any medium regressions for any refactorings of the internals here that might be using different code paths than they used to be: @nanosoldier `runbenchmarks(ALL, vs = ":master")`

(also question for @jrevels, does nanosoldier use a different branch of basebenchmarks for testing against different branches of Julia?)