@Sacha0 2016-12-31
Fixed the ambiguities introduced by rebasing post #19690-merge (and in the appropriate commits). With benefit of #19690, added a separate commit that removes a few vectorized methods over `SparseVector`s and performs minor associated cleanup. Also added a commit replacing removed error-path tests for deprecated vectorized `&`, `|`, `xor`, `min`, and `max` methods over `SparseMatrixCSC`s with similar tests of sparse `map` as discussed in https://github.com/JuliaLang/julia/pull/19713#discussion_r94269292. Thanks again!