@nanosoldier 2016-12-31
[Your benchmark job](https://github.com/JuliaLang/julia/pull/19791#issuecomment-269850951) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/9f35457dae975c90e26cb05b52be80ea4f0fc23a/94f7b60_vs_26c8d85/report.md). cc @jrevels