@tkelman 2016-12-31
That's less change than I expected. string join is often noisy, but `["array","comprehension",("collect","Array{Float64,1}")]` might be real? Let's go again on the updated version @nanosoldier `runbenchmarks(ALL, vs = ":master")`