@nanosoldier 2016-12-31
[Your benchmark job](https://github.com/JuliaLang/julia/pull/19791#issuecomment-269857635) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/7eb72c9e1653485630c558953980fb8e4e67a732/267d9e5_vs_26c8d85/report.md). cc @jrevels