@tkelman 2016-12-31
This is such a conflict-magnet that I'm going to merge now and open my trivial little PR against master instead.