@jrevels 2016-12-31
> (also question for @jrevels, does nanosoldier use a different branch of basebenchmarks for testing against different branches of Julia?)

No, all versions of Julia benchmark against the `nanosoldier` branch.