@giordano 2017-01-03
Going from the old manually vectorized method of the 2-arg variant of `trunc` to the dot syntax, I experience a large performance regression.

Before this PR (manually vectorized method):
```julia
julia> versioninfo()
Julia Version 0.6.0-dev.1843
Commit c38a5a304 (2017-01-01 14:09 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: Intel(R) Core(TM) i7-4700MQ CPU @ 2.40GHz
  WORD_SIZE: 64
  BLAS: libopenblas (USE64BITINT DYNAMIC_ARCH NO_AFFINITY Haswell)
  LAPACK: libopenblas64_
  LIBM: libopenlibm
  LLVM: libLLVM-3.9.1 (ORCJIT, haswell)

julia> using BenchmarkTools

julia> arr = collect(0.0:0.1:10.0);

julia> @benchmark trunc(Int, arr)
BenchmarkTools.Trial: 
  memory estimate:  912.00 bytes
  allocs estimate:  2
  --------------
  minimum time:     368.604 ns (0.00% GC)
  median time:      379.300 ns (0.00% GC)
  mean time:        424.871 ns (6.22% GC)
  maximum time:     4.681 μs (83.10% GC)
  --------------
  samples:          10000
  evals/sample:     207
  time tolerance:   5.00%
  memory tolerance: 1.00%
```
after the PR (dot syntax):
```julia
julia> versioninfo()
Julia Version 0.6.0-dev.1807
Commit 26c8d856a (2016-12-31 04:12 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: Intel(R) Core(TM) i7-4700MQ CPU @ 2.40GHz
  WORD_SIZE: 64
  BLAS: libopenblas (USE64BITINT DYNAMIC_ARCH NO_AFFINITY Haswell)
  LAPACK: libopenblas64_
  LIBM: libopenlibm
  LLVM: libLLVM-3.9.1 (ORCJIT, haswell)

julia> using BenchmarkTools

julia> arr = collect(0.0:0.1:10.0);

julia> @benchmark trunc.(Int, arr)
BenchmarkTools.Trial: 
  memory estimate:  3.61 kb
  allocs estimate:  132
  --------------
  minimum time:     46.124 μs (0.00% GC)
  median time:      47.363 μs (0.00% GC)
  mean time:        48.439 μs (0.74% GC)
  maximum time:     3.685 ms (97.86% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

Can someone please confirm?