@KristofferC 2017-01-03
Make sure to not benchmark in global scope and use `const` but even so, there is a significant regression.