@giordano 2017-01-03
Thanks, I updated my benchmarks with the use of `const`, `trunc` is still 100× slower after this PR.  I'm going open a new issue.