@tkelman 2017-01-05
We somehow didn't see this in any tests, but ref https://discourse.julialang.org/t/is-0-a-1-going-away/1313/2 - chained comparisons like `A .< B .< C` expand to `(A .< B) & (B .< C)`, which now throws a depwarn if any are arrays. Would it be always equivalent to the former behavior if we made it instead use `.&` in the expansion, or would it have to depend on the types? cc @stevengj