@stevengj 2017-01-05
It should be equivalent to the former behavior if we made it use `.&`.