@Sacha0 2017-01-10
Note that the above comparison isn't apples-to-apples: sparse `map[!]/broadcast[!]` avoids storing numerical zeros, whereas the specialization above does not. (The above performance difference is in line with that observed between the former `nz2z` and `nz2nz` sparse `broadcast` classes for similar cases, as benchmarked/discussed elsewhere.) Best!